anomaly-service/pyproject.toml
[project]
name = "anomaly-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi==0.115.0",
    "httpx==0.27.0",
    "numpy==1.26.4",
    "pydantic==2.7.0",
    "python-dateutil==2.9.0",
    "scipy==1.13.1",
    "uvicorn[standard]==0.30.0",
]
